pyproject.toml
[project]
name = "curated-search"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "celery~=5.3.0",
    "django~=4.2",
    "django-allauth~=0.51.0",
    "django-crispy-forms~=1.14.0",
    "django-debug-toolbar~=4.2.0",
    "feedparser~=6.0.10",
    "lxml-html-clean>=0.4.1",
    "newspaper4k>=0.9.3.1",
    "psycopg2-binary~=2.9.3",
    "pytest~=7.1.2",
    "pytest-django~=4.5.2",
    "redis~=4.3.4",
    "requests~=2.28.1",
]
